Moved `InputVariants` to a separate file.
